+++ Lilu/Headers/kern_patcher.hpp
@@ -18,8 +18,7 @@
 
 namespace Patch { union All; void deleter(All * NONNULL); }
 #ifdef LILU_KEXTPATCH_SUPPORT
-struct OSKextLoadedKextSummaryHeader;
-struct OSKextLoadedKextSummary;
+union OSKextLoadedKextSummaryHeaderAny;
 #endif /* LILU_KEXTPATCH_SUPPORT */
 
 class KernelPatcher {
@@ -603,7 +602,7 @@ private:
 	/**
 	 *  A pointer to loaded kext information
 	 */
-	OSKextLoadedKextSummaryHeader **loadedKextSummaries {nullptr};
+	OSKextLoadedKextSummaryHeaderAny **loadedKextSummaries {nullptr};
 
 	/**
 	 *  A pointer to kext summaries update
@@ -616,7 +615,7 @@ private:
 	 *  @param summaries loaded kext summaries
 	 *  @param num       number of loaded kext summaries
 	 */
-	void processAlreadyLoadedKexts(OSKextLoadedKextSummary *summaries, size_t num);
+	void processAlreadyLoadedKexts(OSKextLoadedKextSummaryHeaderAny *summaries, size_t num);
 
 #endif /* LILU_KEXTPATCH_SUPPORT */