+++ Lilu/Headers/kern_util.hpp
@@ -327,6 +327,7 @@ enum KernelVersion {
 	HighSierra    = 17,
 	Mojave        = 18,
 	Catalina      = 19,
+	BigSur        = 20,
 };
 
 /**