+++ Changelog.md
@@ -1,5 +1,9 @@
 Lilu Changelog
 ==============
+#### v1.4.6
+- Added preliminary definitions for 11.0 support
+- Temporarily disabled user patcher for 11.0
+
 #### v1.4.5
 - Fixed newer CPU generation detection
 - Added failsafe versions of CML framebuffers